Add algo2

+++ DocumentSelectionAlgorithms/algo1.h
@@ -15,7 +15,7 @@ void run_algo1_for_query(int qid, FILE *ofile);
 //extern void run_algo12_for_query(int qid, FILE *ofile);
 //extern void run_algo13_for_query(int qid, FILE *ofile);
 //extern void run_algo14_for_query(int qid, FILE *ofile);
-extern long double get_ndcg(vector < int > rel);
+extern long double get_ndcg(vector < long double > rel);
 
 //extern const int DOC_SIZE;
 //extern const int QID_SIZE;
